refactor: move selector packages under balancer
