+++ oxiad/coordinator/balancer/selector/ensemble/selector.go
@@ -17,8 +17,8 @@ package ensemble
 import (
 	"github.com/emirpasic/gods/v2/sets/linkedhashset"
 
-	"github.com/oxia-db/oxia/oxiad/coordinator/selector"
-	"github.com/oxia-db/oxia/oxiad/coordinator/selector/single"
+	"github.com/oxia-db/oxia/oxiad/coordinator/balancer/selector"
+	"github.com/oxia-db/oxia/oxiad/coordinator/balancer/selector/single"
 )
 
 var _ selector.Selector[*Context, []string] = &ensemble{}
