+++ oxiad/coordinator/balancer/selector/single/context.go
@@ -20,7 +20,7 @@ import (
 	"github.com/emirpasic/gods/v2/sets/linkedhashset"
 
 	commonproto "github.com/oxia-db/oxia/common/proto"
-	"github.com/oxia-db/oxia/oxiad/coordinator/model"
+	"github.com/oxia-db/oxia/oxiad/coordinator/balancer/model"
 	"github.com/oxia-db/oxia/oxiad/coordinator/util"
 )
