+++ oxiad/coordinator/balancer/selector/single/selector.go
@@ -17,7 +17,7 @@ package single
 import (
 	"github.com/pkg/errors"
 
-	"github.com/oxia-db/oxia/oxiad/coordinator/selector"
+	"github.com/oxia-db/oxia/oxiad/coordinator/balancer/selector"
 )
 
 var _ selector.Selector[*Context, string] = &server{}
